+++ utils/src/test/java/com/leo/utils/StringUtilsTest.java
@@ -1,0 +1,42 @@
+package com.leo.utils;
+
+import org.junit.Test;
+
+import java.util.Locale;
+
+import static org.junit.Assert.*;
+
+/**
+ * @author Leo.ZhangTJ
+ * @time 2019/4/10
+ * @function
+ * @describe
+ */
+public class StringUtilsTest {
+
+    @Test
+    public void stringAppend() {
+    }
+
+    @Test
+    public void stringAppend1() {
+    }
+
+    @Test
+    public void isConSpeCharacters() {
+    }
+
+    @Test
+    public void hidePhone() {
+    }
+
+    @Test
+    public void currencyFormat() {
+        System.out.println(StringUtils.currencyFormat(2000000.45658, Locale.SIMPLIFIED_CHINESE));
+    }
+
+    @Test
+    public void currencyFormat1() {
+
+    }
+}